requirements.txt
asgiref==3.3.1
boto3==1.17.7
botocore==1.20.7
cycler==0.10.0
dj-database-url==0.5.0
Django==3.1.5
django-cleanup==5.1.0
django-crispy-forms==1.11.0
django-filter==2.4.0
django-heroku==0.3.1
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
kiwisolver==1.3.1
matplotlib==3.2.0
numpy==1.20.0
Pillow==8.1.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
